chore: use .js for other platforms
