--- src/reanimated2/NativeReanimated.web.js
@@ -1,3 +1,0 @@
-import reanimatedJS from './js-reanimated';
-
-export default reanimatedJS;
